+++ NEWS.md
@@ -1,3 +1,9 @@
+# InSiGHTS 0.5
+
+* Helper functions to allow creating a derivate range from a given variable `create_derivate_range()`
+* Function to 'clamp' a given raster to a value range via `st_clamp()`.
+* Parameter in `insights_fraction()` to automatically clamp.
+
 # InSiGHTS 0.4
 
 * Bug fixes and more support for future clipping when `stars` is provided.